refactor: 모듈 구조 리팩토링
