+++ build.gradle.kts
@@ -1,0 +1,3 @@
+allprojects{
+    version = System.getenv("VERSION")
+}
