+++ ssu-commerce-cores/ssu-commerce-core-jpa/src/main/kotlin/com/ssu/commerce/core/jpa/BaseEntity.kt
