+++ ssu-commerce-cores/ssu-commerce-core-jpa/src/test/kotlin/com/ssu/commerce/core/CoreApplication.kt
