+++ ssu-commerce-cores/ssu-commerce-core-jpa/src/test/kotlin/com/ssu/commerce/core/test/CoreApplicationTests.kt
@@ -1,0 +1,12 @@
+package com.ssu.commerce.core.test
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest(properties = ["spring.profiles.active:test"])
+class CoreApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+}
