+++ ssu-commerce-cores/ssu-commerce-core-security/src/main/kotlin/com/ssu/commerce/core/security/config/SecurityConfig.kt
@@ -1,0 +1,38 @@
+package com.ssu.commerce.core.security.config
+
+import com.ssu.commerce.core.security.filter.JwtAuthenticationFilter
+import org.springframework.context.annotation.Configuration
+import org.springframework.security.config.annotation.web.builders.HttpSecurity
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter
+import org.springframework.security.config.annotation.web.configurers.ExpressionUrlAuthorizationConfigurer
+import org.springframework.security.config.http.SessionCreationPolicy
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
+
+@Configuration
+class SecurityConfig(
+    private val jwtAuthenticationFilter: JwtAuthenticationFilter,
+    private val urlPermission: UrlPermissionFilter
+) : WebSecurityConfigurerAdapter() {
+    override fun configure(http: HttpSecurity) {
+        http.httpBasic().disable()
+            .csrf().disable()
+            .cors().and()
+            .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+            .and()
+            .authorizeRequests()
+            .antMatchers("/swagger-ui/**").permitAll()
+            .antMatchers("/v3/api-docs/**").permitAll()
+            .apply { urlPermission.urlPermissions(this) }
+            .and()
+            .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter::class.java)
+    }
+}
+
+interface UrlPermissionFilter {
+    fun urlPermissions(authorizeRequests: ExpressionUrlAuthorizationConfigurer<HttpSecurity>.ExpressionInterceptUrlRegistry): ExpressionUrlAuthorizationConfigurer<HttpSecurity>.ExpressionInterceptUrlRegistry =
+        authorizeRequests
+            .anyRequest().authenticated()
+}
+
+@Configuration
+class DefaultUrlPermissionFilter : UrlPermissionFilter
