+++ ssu-commerce-cores/ssu-commerce-core-security/src/main/kotlin/com/ssu/commerce/core/security/jwt/JwtProviderProperties.kt
@@ -1,0 +1,14 @@
+package com.ssu.commerce.core.security.jwt
+
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.boot.context.properties.ConfigurationProperties
+import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.PropertySource
+
+@Configuration
+@PropertySource("classpath:/security/ssu-commerce-jwt.yml")
+@ConfigurationProperties(prefix = "jwt")
+class JwtProviderProperties {
+    @Value("\${secret}")
+    lateinit var secret: String
+}
