+++ ssu-commerce-cores/ssu-commerce-core-security/src/main/kotlin/com/ssu/commerce/core/security/user/UserRole.kt
@@ -1,0 +1,5 @@
+package com.ssu.commerce.core.security.user
+
+enum class UserRole {
+    ROLE_ADMIN, ROLE_USER, ROLE_ANONYMOUS, ROLE_SERVER
+}
