+++ ssu-commerce-cores/ssu-commerce-core-security/src/main/resources/security/ssu-commerce-jwt.yml
