+++ ssu-commerce-cores/ssu-commerce-core-test/build.gradle.kts
@@ -1,0 +1,3 @@
+dependencies {
+    testImplementation("org.springframework.boot:spring-boot-starter-test")
+}
