--- ssu-commerce-cores/ssu-commerce-core-web/src/main/kotlin/com/ssu/commerce/core/configs/DataSourceConfig.kt
@@ -1,31 +1,0 @@
-package com.ssu.commerce.core.configs
-
-import org.slf4j.LoggerFactory
-import org.springframework.boot.jdbc.DataSourceBuilder
-import org.springframework.context.annotation.Bean
-import javax.sql.DataSource
-
-open class DataSourceConfig(
-    val dataSourceProperties: AbstractDataSourceProperties
-) {
-    @Bean
-    open fun dataSource(): DataSource =
-        DataSourceBuilder.create()
-            .driverClassName(dataSourceProperties.driverClassName)
-            .url(dataSourceProperties.dataSource)
-            .username(dataSourceProperties.userId)
-            .password(dataSourceProperties.password)
-            .build()
-            .also {
-                LoggerFactory.getLogger(DataSourceConfig::class.java)
-                    .info("\uD83D\uDCC2 SSUCommerce ${dataSourceProperties.projectName} DataSource 접속이 되었습니다.")
-            }
-}
-
-interface AbstractDataSourceProperties {
-    var projectName: String
-    var dataSource: String
-    var userId: String
-    var password: String
-    var driverClassName: String
-}
