--- ssu-commerce-cores/ssu-commerce-core-web/src/main/kotlin/com/ssu/commerce/core/configs/WebMvcConfig.kt
@@ -1,26 +1,0 @@
-package com.ssu.commerce.core.configs
-
-import com.ssu.commerce.core.security.UserInfoArgumentResolver
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.method.support.HandlerMethodArgumentResolver
-import org.springframework.web.servlet.config.annotation.CorsRegistry
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-
-@Configuration
-class WebMvcConfig(
-    private val userInfoArgumentResolver: UserInfoArgumentResolver
-) : WebMvcConfigurer {
-
-    override fun addCorsMappings(registry: CorsRegistry) {
-        registry.addMapping("/**")
-            .allowedOrigins(
-                "http://localhost:3000",
-                "https://ssu-commerce.github.io"
-            )
-            .allowedMethods("*")
-    }
-
-    override fun addArgumentResolvers(resolvers: MutableList<HandlerMethodArgumentResolver>) {
-        resolvers.add(userInfoArgumentResolver)
-    }
-}
