--- ssu-commerce-cores/ssu-commerce-core-web/src/main/kotlin/com/ssu/commerce/core/security/UserRole.kt
@@ -1,5 +1,0 @@
-package com.ssu.commerce.core.security
-
-enum class UserRole {
-    ROLE_ADMIN, ROLE_USER, ROLE_ANONYMOUS, ROLE_SERVER
-}
