+++ ssu-commerce-cores/ssu-commerce-core-web/src/main/resources/application.yml
@@ -1,5 +1,4 @@
 springdoc:
   swagger-ui:
-    path: /swagger-ui.html
     base-package: "com.ssu.commerce"
     title: Core
