--- ssu-commerce-cores/ssu-commerce-core-web/src/test/kotlin/com/ssu/commerce/core/CoreApplicationTests.kt
@@ -1,12 +1,0 @@
-package com.ssu.commerce.core
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest(properties = ["spring.profiles.active:test"])
-class CoreApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-}
